oldsiva: rename package to legacysiva
