+++ legacysiva/library_test.go
@@ -1,4 +1,4 @@
-package oldsiva
+package legacysiva
 
 import (
 	"path/filepath"
